+++ packages/i18nify-playground/src/sections/number/view/index.js
@@ -1,1 +1,0 @@
-export { default as NumberView } from './number-view';